chore: Update catalog code to new code structure
Update catalog api and backend to new code structure that separates more
code into different files and out of the mod file.
