+++ src/catalog/backends.rs
@@ -12,5 +12,56 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
+use async_trait::async_trait;
+use dyn_clone::DynClone;
+
 pub mod error;
 pub mod sql;
+
+use crate::catalog::error::CatalogProviderError;
+use crate::catalog::{Endpoint, EndpointListParameters, Service, ServiceListParameters};
+use crate::config::Config;
+use crate::keystone::ServiceState;
+
+#[async_trait]
+pub trait CatalogBackend: DynClone + Send + Sync + std::fmt::Debug {
+    /// Set config
+    fn set_config(&mut self, config: Config);
+
+    /// List services
+    async fn list_services(
+        &self,
+        state: &ServiceState,
+        params: &ServiceListParameters,
+    ) -> Result<Vec<Service>, CatalogProviderError>;
+
+    /// Get single service by ID
+    async fn get_service<'a>(
+        &self,
+        state: &ServiceState,
+        id: &'a str,
+    ) -> Result<Option<Service>, CatalogProviderError>;
+
+    /// List Endpoints
+    async fn list_endpoints(
+        &self,
+        state: &ServiceState,
+        params: &EndpointListParameters,
+    ) -> Result<Vec<Endpoint>, CatalogProviderError>;
+
+    /// Get single endpoint by ID
+    async fn get_endpoint<'a>(
+        &self,
+        state: &ServiceState,
+        id: &'a str,
+    ) -> Result<Option<Endpoint>, CatalogProviderError>;
+
+    /// Get Catalog (Services with Endpoints)
+    async fn get_catalog(
+        &self,
+        state: &ServiceState,
+        enabled: bool,
+    ) -> Result<Vec<(Service, Vec<Endpoint>)>, CatalogProviderError>;
+}
+
+dyn_clone::clone_trait_object!(CatalogBackend);
