+++ src/plugin_manager.rs
@@ -15,7 +15,7 @@
 use std::collections::HashMap;
 
 use crate::assignment::types::AssignmentBackend;
-use crate::catalog::types::CatalogBackend;
+use crate::catalog::backends::CatalogBackend;
 use crate::federation::types::FederationBackend;
 use crate::identity::types::IdentityBackend;
 use crate::resource::types::ResourceBackend;
